Update MainPacman.pde
There is a single variable that has an incorrect value and it ends up breaking our code entirely. When we made the demo file, we had that variable set to a different value for testing purposes, then made the demo before remembering to fix it. When we did try to commit and fix it, it ended up happening to the main file and not the demo file. I am fixing that mistake now.